disable open.in.view for local compose
I don't really understand this, but from what I can glean this property is set to true by default
and causes a transaction to be made and kept open for each controller mapping.
The transaction would only end when a response is returned.
I believe it is defaulted to true for use with the embedded tomcat server,
which I am using during dev. During deployment, even local deployment, it should probably be set to false.