defined Mother concept
This change is meant to lower the friction to testing. What I've noticed
is that the quantity and quality of tests I write are really dependant
on the DX of writing the darn test.

In my head, tests are always adding value. The thing that keeps me from
writing tests is usually creating test data, or setting up the
conditions of the situation I am trying to test.

I imagine defining a Mother for each Domain object that can be easily
interacted with to create dummy data will be a good way to lower that
friction.