added Mother#get(int) convenience
Rather than require every Mother to implement the same stream/iteration
logic, it can be defined by default by the Mother. In my head, this is a
similar idea to a Strategy pattern, just flipped on its head. The Mother
is defining a contract that must be fulfiled (transitively,
Supplier#get) and then building behaviour based on the implementation of
that contract.

[housekeeping] prefer Collections.emptySet() to List.of()