defined Controller annotation
The purpose of this annotation is to document the expectation of each
Controller. Requiring 'resource' and 'respondsWith' is a soft way to
enforce the design I am trying to use here.

In the far future, if the annotation is still in use, the metadata could
be used to write an architecture test.