standardised GET playlists/{id}/songs
As I get closer to the end of this push, the need for an
ExceptionHandler and a custom assertion like 'AssertedResourceNotFound'
is becoming more aparent.

That, and perhaps a single point of reference for each Controller (a Port).
I am considering bringing back the ControllerPort concept, but now for a
better reason.

[tests]
* added PlaylistSongsController#GetSongs test.
** thereAreSongs
** thereAreNoSongs
** assertedPlaylistDoesNotExist