extracted DomainEntity
This change serves to standardise how I want to write entities in the
Domain, and a place to put documentation on the reasoning.

I realise that before I can standardise the last PlaylistsSongs
endpoints, I really do need a technical identifier to address the Song
in a Playlist to perform actions like Delete, etc.

[revisions]
* added technical identifier to Song

[additions]
* added EntityMother extension to Mother