defined AssertedResourceNotFoundException
While this change does reduce a few lines from each Controller, and
denest certain methods, it is superflous. Those execution paths are
still present, just handled indirectly.

The purpose of this change is to be a first step towards standardising the
reporting of errors and exceptions (problems) on the API. Rather than explicitly
handling this in every Controller method, we want to delegate it to a
single ExceptionHandler.

This won't be done for all problems, only ones that 'cross-cutting'.

[additions]
* added ControllerExceptionHandler to define the ResponseEntities for
  various custom Exceptions we will define.

[revisions]
* update Controllers to leverage the new Exception.