revise QueryRepository
I am still playing with the idea of re-introducing Ports, specifically a
ResourcePort. However, the main thing I need is just a Resource#get(ID)
signature - which the QueryRepository can provide. In fact, it makes
sense for the 'resource' layer to handle this concern.

The ExceptionHandler can stay as is, with 'web' specific details being
built around the AssertedResourceNotFound exception, rather than the
Exception encapsulating that.

[revisions]
* AssertedResourceNotFoundException now belongs to 'resource.query'
  package.