added SongService
My mistake with PlaylistService was to define the API before I even
began to model the RESTful API, or consider the domain.

The driver of the development should be the needs of the consumers of
the RESTful API, and then the modelling of the domain. The service
should evolve to meet those needs.

[revisions]
* deprecate ServiceImpl#retrieve. This was a defensive coding artefact
  that is not necessary in this codebase.