extracted PlaylistSong
This resource existed, but was mapped using ORM magic and 'hidden' away
in the Playlist#songs Collection. A PlaylistSong is a key domain
concept, it is where we will store Order, Favouriting, Preferred
Provider, etc.

The main reason for the extraction it support the needs of the API; i.e.
serving PlaylistSongs and updating their Order, etc.

[revisions]
* marked Playlist#songs usages as deprecated. They will be phased out in
  favour of the PlaylistSong abstraction, and then removed.

[additions]
* extracted CreateResource interface. May roll it back, but for now the
  general contract for this stands. It would be the same across the rest
  of the CRUD Services. Since not all services will expose Update or
  Delete, it is a slim interface mainly for documentation.