redefined ResourceNotFoundException
The need for the Exception was really a 'web' module concern that is
better handled by using exisiting Optional signatures, which will always
be more flexible.

Rather than a CheckedException to force a 'throws' declaration, we can
just use '.orElseThrow(...)' to make the code self-document the flow -
as it should.

[revisions]
* deprecates AssertedResourceNotFoundException.
* prefer 404 Not Found with a ProblemDetail for further clarity, rather
  than confusing things with 400 Bad Request.

[housekeeping]
* resolves TODO to return ProblemDetail for the Exception.