added #verifyUrl to handle 3xx redirections
This first pass proves the concept, but there are issues with the way
this is implemented. There needs to be a handler for calling this method
because we may end up in a situation where a URL is redirected to a URL
of a Song we already store. In such a case we need to delete one of
them, but if any PlaylistSong(s) point to one, they must be migrated.