rm OEmbedService
This change moves the lookup logic to the Provider class itself,
eliminating an unneccessary middleman. This change also means we
can flatten the OEmbed module into the Core module.

The reason for this change is because co-ordination between Song,
PlaylistSong and Provider is needed during lookup, so an OEmbedService
is not the correct service model.

The scheduling of the job to lookup should also be an Application module
concern, with the other modules simply exporting their libraries*

[removals]
* rm #lookupSong scheduled job. This operation requires an actual
  service to handle co-ordination between Song, PlaylistSong and
  Provider.