explicitly define DDL
[fixes] varchar(255) being used for URLs. Though this works for most
cases, the actual maximum length of a URL, though not standardised, can
be around 2,000 characters. This caused an issue recently when the
system tried to handle a redirection URL from Deezer that mapped to a
location whose URL included many query parameters and whose lenght
exceeded the default varchar(255) used by Hibernate.

[revisions]
* specify the database name in `POSTGRES_DB` environment variable accepted
  by the Postgres image, rather than messing around with a 'create
  database gm_song_storage & \c gm_song_storage' in the DDL script.
* use `spring.jpa.hibernate.ddl-auto=update` to allow Hibernate to
  handle the creation of the Sequence tables per JPA specification. I do
  not want to mess around with defining the sequences myself as yet.