[fix] #verifyUrl incorrectly handling Song#lastLookup
This issue came about because I am not managing the state of Song correctly.
While property based injection (JPA calling the accessor/mutator methods) is
good to reason about, it does constrain how I model my Domain Objects - which
is not good.

If I were to add validation and invariant management logic in the mutators,
they would be invoked when JPA is trying to map the entity into an object after
retrieving it from the DB.

I need to switch to field based injection.

[revisions]
* amend SongLookupService logging for more clarity on what the system is
  doing.