rev PlaylistSong#<init>
A PlaylistSong can serve as a creation root for a Song. Currently, this
is the primary mode of creation for a Song.

It is good that we experimented with different styles here. We needed to
know how we are going to handle bridging entities and now we do. There
is an owning side that must exist, and the bridging entity can act as a
creation root for the owned side.

[housekeeping]
- mv PlaylistSongsCreateRequest SongsCreateRequest. Incorrect naming
  convention for what it actually was creating.