rm resource.query layer (#6)
The idea was worth exploring, and may have a use case in a different
domain/system, but in this one it just introduces more files to juggle.
It also made test-setup/mocking a little more tedious in testing.

Less is still more.

[removals]
- rm "_ResourceMarker". Flattening of "resource" package now gives us a
target, "EntityRepository", to perform entity scanning from; we no
longer need the marker.

[revisions]
- "service" layer is now the only layer that interfaces with the
"resource" layer.

[additions]
- added "Read" semantic, mv "CreateResource" "Create" for consistency.

[housekeeping]
- mv "service", "repository" field names to collective nouns, e.g.
"songs", "playlists", etc. Reads better to me, and encourages naming
methods as if they were operations to the collection.
- rm query semantics from "service" method names.