added UnitTest, IntegrationTest tags (#9)
Added definitions, in the style of SmokeTest, to UnitTest and
IntegrationTest semantic annotations. These tests are tagged as well,
with corresponding alias rules provided in the Makefile.

Being able to control the slice of tests being run is already extremely
useful. I believe this will pay dividends as the system grows.

[fixes]
- missing UnitTest annotation on DomainEntityTest.

[revisions]
- mv *ControllerTest *ControllerIT. They are slice tests spinning up
partial context.
- revert pre-commit hook; add back UnitTest check.
- mv SmokeTest check to pre-push hook; rm IntegrationTest check. Perhaps
in a later iteration we can run the Integration tests are a pre/post
merge hook, if that is possible.

[housekeeping]
- amend `make help` command "column" width to fit new rule target.