collapse *ServiceImpl into *Service (#10)
While programming to the interface is a good default mode, in this
system we are gaining nothing from this thin layer of abstraction. We
have no services, currently, that have or need multiple implementations.
For testing, we are leveraging Mockito and not implementing a stub of
the interface.

We can control the service api via considered use of the `public`
modifier, we don't need an interface. If a real need for an interface
arises, it would be trivial for us to introduce one. For now, less is
more.