core packaging: mv layer-based domain-based (#11)
Layer-based is fine, and it can still work (`web` module), but there is
a need to leverage default(package-private) access in the application
code.

`PlaylistSong` needs to expose a method to migrate the `Song` it points
to, but this requires coordination. It has to happen through a service
layer. By keeping the service definition within the same package, the
service can expose the migration method and call into the
package-private migration method handling the coordination.

[fixes]
- access modifiers on PlaylistSong: mv `protected`
default(package-private).

[revisions]
- mark Song#lookup package-private. It is now exposed through the
LookupService, which provides the coordination the method requires.
- mark Provider package-private. We can consider this an implementation
detail for Song lookups.
- CoreContext must scan from itself recurisvely, as the targets are no
longer all grouped in a single (layer) package.
- *Mother#setId classes must now delegate to EntityMother#setId as
DomainEntity is now in the `semantic` package, and #setId is
package-private.

[additions]
- added CoreService annotation to mark service layer. EntityRepository
can mark resource layer, and the `core` module itself is the domain
layer.
- added AccessForTesting annotation to markup methods whose access was
widened to enable/support testing.

[housekeeping]
- mv package-info documentation into layer marker classes.